Make ready for multiple engine usage (#7)
With this update GEANT4_VMC becomes capable of running together with
other engines sharing the simulation of one event. Depending on user
conditions, tracks might be transported by an engine based on geometry
constraints, particle type or phase space. When these conditions are not
met anymore, the ROOT TMCManager interrupts the transport and transfers
the track to the engine responsible under the new conditions. This can
happen multiple times and is one automatically until all tracks have
been fully transported.

The new example E03c demonstrates this by splitting the simulation of a
sampling calorimeter. The active (passive) layers are simulated with
GEANT3 (GEANT4) and the passive (active) layers with GEANT4 (GEANT3).