Fix in TGeant4::Mixture(Int_t& kmat, const char* name, Float_t* ...):
When using geometry defined via Root or via VMC with Root navigation,
this function is delegated to TGeoMCGeometry class defined in Root/vmc, where
'wmat' values were not updated when the function is called with nlmat<0, as they
should be to mimic the Geant3 behavior.
The bug in Root is fixed in the patches for versions 6.18/06 and 6.20/00,
the older versions are fixed with this work-around.