Faster TGeo<->G4 geometry lookups
Change of type map --> unordered_map for the G4 <-> TGeo
translation.
This is advantageous since the structures are read-only (random key access)
and so we go from O(log(N)) to O(1) scaling on average.

In tests for the ALICE detector, pp events, this commit results
in a noticable speed increase (~5%) in the overall G4 runtime.