Allow to invoke random seed propagations to Geant4 (#45)
So far, we can only propagate or pick up the seed
from gRandom to Geant4 (CLHEP) during initialization.

This commit makes minimal changes that allows an expert
user to call TGeant4::SetRandomSeed() at any moment,
thereby gaining better control over the random state of
the Geant4 engine. The call needs to be passed via TGeant4::SetRandomSeed
so that we don't need to expose TG4RunManager in ROOT dictionaries.

Motivation:

In the multi-process + sub-event parallel framework of ALICE,
we are in the need of an interface that allows us to set the
random state during simulation runs, in order to
ensure total reproducibility of our simulations with Geant4 when
starting from the same initial seed. For Geant3, this was
already working but reproducibility was so far lacking for Geant4.

The present change (or similar) is also need for the even higher
goal of "track seeding", which gives reproducible simulations on
the track level even when cuts or geometry change in some parts
of the detector.